requirements.txt
tkinter
pandas
requests
pillow
